发布任务界面完善 发布任务接口对接完成 增加查看任务+接口对接 增加修改任务+接口对接

+++ app/src/main/java/xin/lrvik/taskcicleandroid/common/UserInfo.kt
@@ -7,5 +7,5 @@ package xin.lrvik.taskcicleandroid.common
 object UserInfo {
     var longitude: Double = 0.0
     var latitude: Double = 0.0
-    var userId: Int = 6
+    var userId: Long = 6
 }
